Split persistence JSON parser
